refactor: 수업 관리 모달 시간 포맷 변경 & 시간 포맷 함수 common으로 이동
