+++ src/views/admin/util/date.js
@@ -5,3 +5,15 @@ export function extractDateInKorean(dateStr) {
   const day = date.getDate()
   return `${month}월 ${day}일`
 }
+
+export function formatDateTime(isoString) {
+  const date = new Date(isoString);
+
+  const year = date.getFullYear();
+  const month = String(date.getMonth() + 1).padStart(2, '0');
+  const day = String(date.getDate()).padStart(2, '0');
+  const hours = String(date.getHours()).padStart(2, '0');
+  const minutes = String(date.getMinutes()).padStart(2, '0');
+
+  return `${year}년 ${month}월 ${day}일 ${hours}시 ${minutes}분`;
+}
